Process the chromosomes sequentially to avoid to have them interleaved in the simulated SBI. The their order in the command line will be the order in the ouput SBI.

+++ genotype/src/main/java/org/campagnelab/dl/genotype/tools/SBISimulatorArguments.java
@@ -28,7 +28,7 @@ public class SBISimulatorArguments implements ToolArguments {
 
 
     @Parameter(names = "--read-N", description = "Read at most N chromosomes from the varmap, then stop.")
-    public long readN=Long.MAX_VALUE;
+    public int readN = Integer.MAX_VALUE;
 
     @Parameter(names ={"-v", "--verbose"}, description = "Be more verbose.")
     boolean verbose;
